Use C++ header suffix
Also fix implicit conversion issue

+++ kata_2/is_prime_number.cpp
@@ -1,11 +1,11 @@
 #include <cmath>
-#include "is_prime_number.h"
+#include "is_prime_number.hpp"
 
 bool isPrime(int number) {
     if (number < 2) {
         return false;
     }
-    unsigned short sqrt = std::sqrt(number);
+    double sqrt = std::sqrt(number);
     for (int i {2}; i <= sqrt; ++i) {
         if (number % i == 0) {
             return false;
